feat(make old one become two component): make old one become two componet
