+++ src/app/page.tsx
@@ -1,12 +1,15 @@
 'use client';
 
-import { StatusChanging } from '@/components/elements/status';
-import { Button } from '@/components/ui/button';
+import StatusButton from '@/components/elements/status-botton';
+import StatusLabel from '@/components/elements/status-label';
 
 export default function Home() {
   return (
     <div className="min-w-full min-h-full flex flex-col items-center justify-center mt-10 gap-10">
-      <StatusChanging />
+      <div className="flex items-center gap-[17px]">
+        <StatusLabel />
+        <StatusButton />
+      </div>
     </div>
   );
 }
