+++ src/lib/shared.ts
@@ -5,3 +5,9 @@ export interface CommentBoxProp {
   authorId: string;
   createdAt: Date;
 }
+
+export interface Status {
+  value: string;
+  label: string;
+  icon: string;
+}
